+++ table.py
@@ -5,8 +5,10 @@ import csv
 from scraper import kolesar_scraper
 
 
+
 def podatki_v_csv(directory, ime_datoteke, glavni_list):
-    os.remove(directory + "/" + ime_datoteke)
+    if os.path.exists(directory + "/" + ime_datoteke):
+        os.remove(directory + "/" + ime_datoteke)
     fieldnames = list((glavni_list[0]).keys())
     os.makedirs(directory, exist_ok=True)
     path = os.path.join(directory, ime_datoteke)
